Modify UT to make weight constant
